Create new folder to reorganize files
