+++ src/Components/App/App.js
@@ -2,8 +2,8 @@ import React from 'react';
 import { Routes, Route } from 'react-router-dom';
 import './App.css';
 import MainDisplay from './MainDisplay/MainDisplay';
-import DashboardPage from './DashboardPage';
-import DetailsPage from './DetailsPage';
+import DashboardPage from './DashboardPage/DashboardPage';
+import DetailsPage from './DetailsPage/DetailsPage';
 import ListingForm from '../ListingForm/ListingForm';
 
 function App() {
